Bug-Fix: should NOT clear need_memory_commit after first commit_memory_again
We should NOT clear need_memory_commit after we call tm_commit_memory_again for
the first time because we may fail to enter guest this time and we need to
continue to commit memory again and again before we enter guest.
This bug will make some errors after recording for a long time.
This bug is introduced in commit:
TEST: Remove tm_walk_mmu in tm_commit_memory_again

Signed-off-by: Le Tan <tamlokveer@gmail.com>