check and fix spte of rr_cow_page before accessing it
Because spte may be dropped accidentally by mmu notifier, we need to
check and fix the rr_cow_page before accessing it. If the spte has been
dropped, we will call tdp_page_fault to emulate a page access to setup
the spte and fix the rr_cow_page.
When we rollback a cow page, we need to check if that spte has been
dropped, too. But here we will simply give up setting back the original
pfn if it has been dropped.

Signed-off-by: Le Tan <tamlokveer@gmail.com>