change the method to calculate the age of holding_pages
Add rr_vcpu_info.nr_chunk to record current number of chunks. When a
cow_page is added into holding_pages, it will set its chunk_num to
current nr_chunk. When the chunk_num of a cow_page is less than a
certain number, we consider it an old page and release it.

Signed-off-by: Le Tan <tamlokveer@gmail.com>